# docs/openapi/phase-1.yaml
...
openapi: 3.1.0
info:
  title: UB — Phase 1 API
  description: Identity, tenancy, and catalog backbone (Phase 1).
  version: 0.0.1-SNAPSHOT
servers:
- url: /
  description: Current host
paths:
  /api/v1/pricing/price-rules/{ruleId}:
    put:
      tags:
      - pricing-controller
      operationId: updateRule
      parameters:
      - name: ruleId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutPriceRuleRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PriceRuleResponse'
...
  /api/v1/purchasing/path-b/sessions:
    post:
      tags:
      - path-b-purchasing-controller
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePathBSessionRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PathBSessionDetailResponse'
  /api/v1/purchasing/path-b/sessions/{sessionId}/post:
    post:
      tags:
      - path-b-purchasing-controller
      operationId: post
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostPathBRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PostPathBResponse'
  /api/v1/purchasing/path-b/sessions/{sessionId}/lines:
    post:
      tags:
      - path-b-purchasing-controller
      operationId: addLine
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPathBLineRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PathBLineResponse'
...
  /api/v1/purchasing/path-a/purchase-orders/{purchaseOrderId}/lines:
    post:
      tags:
      - path-a-purchasing-controller
      operationId: addLine_1
      parameters:
      - name: purchaseOrderId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPathAPurchaseOrderLineRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PathAPurchaseOrderLineResponse'
...
  /api/v1/pricing/tax-rates:
    get:
      tags:
      - pricing-controller
      operationId: listTaxRates
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaxRateResponse'
    post:
      tags:
      - pricing-controller
      operationId: createTaxRate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostTaxRateRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TaxRateResponse'
  /api/v1/pricing/selling-prices:
    post:
      tags:
      - pricing-controller
      operationId: setSellingPrice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostSellingPriceRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SellingPriceResponse'
  /api/v1/pricing/price-rules:
    get:
      tags:
      - pricing-controller
      operationId: listRules
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PriceRuleResponse'
    post:
      tags:
      - pricing-controller
      operationId: createRule
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostPriceRuleRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PriceRuleResponse'
  /api/v1/pricing/buying-prices:
    post:
      tags:
      - pricing-controller
      operationId: setBuyingPrice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostBuyingPriceRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BuyingPriceResponse'
...
  /api/v1/purchasing/path-b/sessions/{sessionId}:
    get:
      tags:
      - path-b-purchasing-controller
      operationId: get_1
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PathBSessionDetailResponse'
...
  /api/v1/pricing/suggest/sell:
    get:
      tags:
      - pricing-controller
      operationId: suggestSell
      parameters:
      - name: itemId
        in: query
        required: true
        schema:
          type: string
      - name: supplierId
        in: query
        required: false
        schema:
          type: string
      - name: branchId
        in: query
        required: false
        schema:
          type: string
      - name: unitCost
        in: query
        required: false
        description: Draft line buying cost for margin suggestion when historical cost is missing or overridden.
        schema:
          type: number
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SellPriceSuggestionResponse'
...
  /api/v1/inventory/stock-take/sessions/{sessionId}:
    get:
      tags:
      - stock-take-controller
      operationId: getSession
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StockTakeSessionResponse'
...
components:
  schemas:
    PutPriceRuleRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        ruleType:
          type: string
          minLength: 1
        paramsJson:
          type: string
        active:
          type: boolean
      required:
      - name
      - paramsJson
      - ruleType
    PriceRuleResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        ruleType:
          type: string
        paramsJson:
          type: string
        active:
          type: boolean
...
    CreatePathBSessionRequest:
      type: object
      properties:
        supplierId:
          type: string
          maxLength: 36
          minLength: 0
        branchId:
          type: string
          maxLength: 36
          minLength: 0
        receivedAt:
          type: string
          format: date-time
        notes:
          type: string
          maxLength: 5000
          minLength: 0
      required:
      - branchId
      - receivedAt
      - supplierId
    PathBLineResponse:
      type: object
      properties:
        id:
          type: string
        sortOrder:
          type: integer
          format: int32
        descriptionText:
          type: string
        amountMoney:
          type: number
        suggestedItemId:
          type: string
        lineStatus:
          type: string
    PathBSessionDetailResponse:
      type: object
      properties:
        id:
          type: string
        supplierId:
          type: string
        branchId:
          type: string
        receivedAt:
          type: string
          format: date-time
        notes:
          type: string
        status:
          type: string
        lines:
          type: array
          items:
            $ref: '#/components/schemas/PathBLineResponse'
    PostPathBLineBreakdown:
      type: object
      properties:
        lineId:
          type: string
          maxLength: 36
          minLength: 0
        itemId:
          type: string
          maxLength: 36
          minLength: 0
        usableQty:
          type: number
        wastageQty:
          type: number
      required:
      - itemId
      - lineId
      - usableQty
      - wastageQty
    PostPathBRequest:
      type: object
      properties:
        lines:
          type: array
          items:
            $ref: '#/components/schemas/PostPathBLineBreakdown'
          minItems: 1
      required:
      - lines
    PostPathBResponse:
      type: object
      properties:
        supplierInvoiceId:
          type: string
        invoiceNumber:
          type: string
        journalEntryId:
          type: string
        grandTotal:
          type: number
        linesPosted:
          type: integer
          format: int32
    AddPathBLineRequest:
      type: object
      properties:
        description:
          type: string
          maxLength: 2000
          minLength: 0
        amountMoney:
          type: number
        suggestedItemId:
          type: string
          maxLength: 36
          minLength: 0
      required:
      - amountMoney
      - description
...
    PathAPurchaseOrderLineResponse:
      type: object
      properties:
        id:
          type: string
        sortOrder:
          type: integer
          format: int32
        itemId:
          type: string
        qtyOrdered:
          type: number
        qtyReceived:
          type: number
        unitEstimatedCost:
          type: number
    AddPathAPurchaseOrderLineRequest:
      type: object
      properties:
        itemId:
          type: string
          minLength: 1
        qtyOrdered:
          type: number
        unitEstimatedCost:
          type: number
      required:
      - itemId
      - qtyOrdered
      - unitEstimatedCost
...
    PostTaxRateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        ratePercent:
          type: number
          minimum: 0
        inclusive:
          type: boolean
        active:
          type: boolean
      required:
      - name
      - ratePercent
    TaxRateResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        ratePercent:
          type: number
        inclusive:
          type: boolean
    PostSellingPriceRequest:
      type: object
      properties:
        itemId:
          type: string
          minLength: 1
        branchId:
          type: string
        price:
          type: number
          minimum: 0.01
        effectiveFrom:
          type: string
          format: date
        notes:
          type: string
      required:
      - effectiveFrom
      - itemId
      - price
    SellingPriceResponse:
      type: object
      properties:
        id:
          type: string
        itemId:
          type: string
        branchId:
          type: string
        price:
          type: number
        effectiveFrom:
          type: string
          format: date
        effectiveTo:
          type: string
          format: date
    PostPriceRuleRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        ruleType:
          type: string
          minLength: 1
        paramsJson:
          type: string
        active:
          type: boolean
      required:
      - name
      - paramsJson
      - ruleType
    PostBuyingPriceRequest:
      type: object
      properties:
        itemId:
          type: string
          minLength: 1
        supplierId:
          type: string
          minLength: 1
        unitCost:
          type: number
          minimum: 1.0E-4
        effectiveFrom:
          type: string
          format: date
        sourceType:
          type: string
        notes:
          type: string
      required:
      - effectiveFrom
      - itemId
      - supplierId
      - unitCost
    BuyingPriceResponse:
      type: object
      properties:
        id:
          type: string
        itemId:
          type: string
        supplierId:
          type: string
        unitCost:
          type: number
        effectiveFrom:
          type: string
          format: date
        effectiveTo:
          type: string
          format: date
...
    StockAdjustmentRequestResponse:
      type: object
      properties:
        id:
          type: string
        stockTakeLineId:
          type: string
        itemId:
          type: string
        varianceQty:
          type: number
        systemQtySnapshot:
          type: number
        countedQty:
          type: number
        status:
          type: string
    StockTakeLineResponse:
      type: object
      properties:
        id:
          type: string
        itemId:
          type: string
        systemQtySnapshot:
          type: number
        countedQty:
          type: number
        note:
          type: string
    StockTakeSessionResponse:
      type: object
      properties:
        id:
          type: string
        branchId:
          type: string
        status:
          type: string
        notes:
          type: string
        closedAt:
          type: string
          format: date-time
        lines:
          type: array
          items:
            $ref: '#/components/schemas/StockTakeLineResponse'
        adjustmentRequests:
          type: array
          items:
            $ref: '#/components/schemas/StockAdjustmentRequestResponse'
...
    SellPriceSuggestionResponse:
      type: object
      properties:
        latestUnitCost:
          type: number
        marginPercent:
          type: number
        ruleName:
          type: string
        suggestedSellPrice:
          type: number
        note:
          type: string
        currentSellPrice:
          type: number
...
  securitySchemes:
    bearer:
      type: http
      description: 'Slice 3: access JWT via Authorization header.'
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      description: 'Slice 3: tenant integration key (prefix + secret; returned once
        at create).'
      name: X-API-Key
      in: header
    cookieRefresh:
      type: apiKey
      description: Optional HttpOnly refresh cookie; native clients may use JSON body
        instead.
      name: ub_refresh
      in: cookie